requirements.txt
python-telegram-bot==21.5
anthropic>=0.50.0
python-dotenv==1.0.1
